feat: update Confetti component
