+++ app/[locale]/components/Confetti/index.ts
@@ -1,0 +1,1 @@
+export { Confetti } from './Confetti'
