+++ package.json
@@ -9,17 +9,18 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@tsparticles/engine": "^3.5.0",
+    "@tsparticles/preset-confetti": "^3.1.0",
+    "@tsparticles/react": "^3.0.0",
     "firebase": "^10.7.0",
     "framer-motion": "^10.16.10",
-    "next": "14.0.1",
+    "next": "^14.2.6",
     "next-intl": "^3.2.1",
     "react": "^18",
     "react-cookie": "^6.1.1",
     "react-dom": "^18",
     "react-loading-skeleton": "^3.3.1",
-    "react-particles": "^2.12.2",
     "sharp": "^0.32.6",
-    "tsparticles": "^2.12.0",
     "use-sound": "^4.0.1"
   },
   "devDependencies": {
@@ -48,7 +49,10 @@
   "release": {
     "branches": [
       "main",
-      { "name": "develop", "prerelease": true }
+      {
+        "name": "develop",
+        "prerelease": true
+      }
     ],
     "prepare": [
       "@semantic-release/changelog",
